delete confirm

+++ src/main/webapp/jsp/article/detail.jsp
@@ -10,18 +10,20 @@
 <html>
 <head>
 <meta charset="UTF-8">
+<link rel="stylesheet" href="../css/style.css" />
 <title>Insert title here</title>
 </head>
 <body>
-	<div style="border: 2px solid black; width: 500px; height: 500px;">
+	<div style="border: 2px solid black; height: 500px;">
 		<div>번호 : <%= (int) todo.get("id") %></div>
 		<div>작성일 : <%= (LocalDateTime) todo.get("regDate") %></div>
 		<div>수정일 : <%= (LocalDateTime) todo.get("updateDate") %></div>
 		<div>제 목 : <%= (String) todo.get("title") %></div>
-		<div>내 용 : <%= (String )todo.get("body") %></div>
+		<div style="white-space: pre-wrap;">내 용 : <%= (String )todo.get("body") %></div>
 	</div>
 	<a href="modify?id=<%=(int)todo.get("id")%>">게시물 수정</a>
-	<a href="delete?id=<%=(int)todo.get("id")%>">게시물 삭제</a>
+	<a href="delete?id=<%=(int)todo.get("id")%>" onclick="return confirm('게시물을 삭제하시겠습니까?')">게시물 삭제</a>
+	<br />
 	<br />
 	<a href="list">게시물 목록</a>
 </body>
